Re-introduce bookmark modal - add/edit current site uses the bookmark hanger - all other cases use a new control which wraps a modal around a slightly different   version of the bookmark hanger.
Fixes https://github.com/brave/browser-laptop/issues/5103

Auditors: @jkup, @bbondy

UPDATE:
modal is now centered and also features a close button :)
